backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.45.0
Pillow==10.4.0
openai-whisper>=20231117
yt-dlp>=2026.3.17
pdfplumber==0.11.4
opencv-python-headless==4.10.0.84
numpy>=1.26.4
httpx>=0.27.2
psutil>=5.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
